Restructured project and accompanying tests
